options/linux: move over linux extensions of sched.h
