+++ options/linux/include/mlibc/linux-sysdeps.hpp
@@ -1,6 +1,7 @@
 #ifndef MLIBC_LINUX_SYSDEPS
 #define MLIBC_LINUX_SYSDEPS
 
+#include <sched.h>
 #include <stdarg.h>
 #include <sys/epoll.h>
 #include <sys/sysinfo.h>
@@ -68,6 +69,8 @@ int sys_write(int fd, const void *buf, size_t count, ssize_t *bytes_written);
 [[gnu::weak]] int sys_removexattr(const char *path, const char *name);
 [[gnu::weak]] int sys_lremovexattr(const char *path, const char *name);
 [[gnu::weak]] int sys_fremovexattr(int fd, const char *name);
+
+[[gnu::weak]] int sys_getaffinity(pid_t pid, size_t cpusetsize, cpu_set_t *mask);
 
 } // namespace mlibc
