+++ options/posix/generic/sched-stubs.cpp
@@ -17,15 +17,6 @@ int sched_yield(void) {
 	return 0;
 }
 
-int sched_getaffinity(pid_t pid, size_t cpusetsize, cpu_set_t *mask) {
-	MLIBC_CHECK_OR_ENOSYS(mlibc::sys_getaffinity, -1);
-	if(int e = mlibc::sys_getaffinity(pid, cpusetsize, mask); e) {
-		errno = e;
-		return -1;
-	}
-	return 0;
-}
-
 int sched_get_priority_max(int) {
 	__ensure(!"Not implemented");
 	__builtin_unreachable();
@@ -61,27 +52,7 @@ int __mlibc_cpu_count(const cpu_set_t *set) {
 	return count;
 }
 
-int unshare(int) {
-	__ensure(!"Not implemented");
-	__builtin_unreachable();
-}
-
-int clone(int (*)(void *), void *, int, void *, ...) {
-	__ensure(!"Not implemented");
-	__builtin_unreachable();
-}
-
 int sched_setscheduler(pid_t, int, const struct sched_param *) {
-	__ensure(!"Not implemented");
-	__builtin_unreachable();
-}
-
-int sched_setaffinity(pid_t, size_t, const cpu_set_t *) {
-	__ensure(!"Not implemented");
-	__builtin_unreachable();
-}
-
-int sched_getscheduler(pid_t) {
 	__ensure(!"Not implemented");
 	__builtin_unreachable();
 }
